server/requirements.txt
alembic==1.7.5
click==8.0.3
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.10.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.29
Werkzeug==2.0.2
zipp==3.7.0
